Added PythonInstall.ps1 to automatically install python. And the Readme file for  instructions on what PythonInstall does. PythonInstall does not only install Python v3.6.7 64 bit to the users computer, but it does it in the background, it sets enviormental vars for pip and python and it installs numpy, SpeechRecognition and PyAudio to the system.
It is 100% crutal that you run this file before testing this on your app, since both SpeechRecognition and PyAudio is required for python to detect your voice and translate that to text.

This file installs python correctly and sets the environmental vars correctly, but it hasn't been tested on a fresh pc if it installs SpeechRecognition or PyAudio correctly. However, this will be tested very shortly and if there is an error, this will be updated.

For instructions on when you should deploy this Powershell script on a client's computer, and how you should do it on non-tech savy users computers, please check the Readme.txt file in the Install Refrences folder. Note : This isn't coppied to the users computer on build so its highly recomended you either run this file during install or on application startup since it will run in the background.